requirements.txt
numpy~=1.26.4
pyodbc~=5.1.0

python-dotenv~=1.0.1